Added missing helper functions; fixed error with max of empty set
