+++ experiments/eccb2016/scripts/sample_mutation_frequency_plot.py
@@ -4,7 +4,7 @@
 import matplotlib
 matplotlib.use('Agg')
 import sys, os, argparse, json, matplotlib.pyplot as plt, numpy as np, seaborn as sns, pandas as pd
-from ragrpyfu.io import aligned_plaintext_table
+from helper import aligned_plaintext_table
 plt.style.use('ggplot')
 
 # Parse required arguments
